feat: add cli args and inference

+++ src/pente_dataloader.py
@@ -6,9 +6,9 @@ from src.game.pente.pente_game import PenteGame
 
 def board_to_cpu_tensor(board: np.ndarray) -> torch.Tensor:
     # TODO: this only supports two players
-    p1_mask = (board == PenteGame.PLAYER_ONE).astype(np.float32)
-    p2_mask = (board == PenteGame.PLAYER_TWO).astype(np.float32)
-    return torch.from_numpy(np.stack([p1_mask, p2_mask], axis=0))
+    p1_mask = (board == PenteGame.PLAYER_ONE)
+    p2_mask = (board == PenteGame.PLAYER_TWO)
+    return torch.from_numpy(np.stack([p1_mask, p2_mask], axis=0)).type(dtype=torch.float32)
 
 class PenteDataset(Dataset):
     def __init__(self, games: list[tuple[np.ndarray, np.ndarray, float]]):
